Shared Element Transition - Complete with bugfix
Transition was not happening when page was changed to a view that was not initially visible in the grid layout, meaning the user did not scroll down, but when to this view with page change. On returning the shared transition is supposed to happen on this view but in the shared element callback this view was null. It was not bound, so i called a notifyitemchanged on the view at the given position to force binding of the view. Also if the viewholder for this view was null, i waited for the view to be attached to the window before calling starttransition